+++ modules/extension/xsd/xsd-core/src/main/java/org/geotools/xs/bindings/XSNonPositiveIntegerBinding.java
@@ -43,6 +43,8 @@ import org.geotools.xs.XS;
  * </p>
  *
  * @generated
+ *
+ * @source $URL: http://svn.osgeo.org/geotools/tags/2.6.2/modules/extension/xsd/xsd-core/src/main/java/org/geotools/xs/bindings/XSNonPositiveIntegerBinding.java $
  */
 public class XSNonPositiveIntegerBinding implements SimpleBinding {
     final BigInteger MIN_LONG = BigInteger.valueOf(Long.MIN_VALUE);
@@ -114,7 +116,7 @@ public class XSNonPositiveIntegerBinding implements SimpleBinding {
     public String encode(Object object, String value) throws Exception {
         Number number = (Number) object;
 
-        if (number.intValue() > 0) {
+        if (number.longValue() > 0) {
             throw new IllegalArgumentException("Value '" + number
                 + "' must be non-positive (0 or below).");
         }
